requirements.txt
lxml>=4.9
Pillow>=9.0

